+++ src/geometries/point.ts
@@ -1,6 +1,6 @@
+import { tileEffectiveHeight, tileEffectiveWidth, tileMatrixToBBox } from '../tiles';
 import { Tile } from '../tiles/tile';
 import type { TileMatrixSet } from '../tiles/tileMatrixSet';
-import { tileEffectiveHeight, tileEffectiveWidth, tileMatrixToBBox } from '../tiles/tiles';
 import type { TileMatrixId } from '../tiles/types';
 import { validateCRSByOtherCRS, validateMetatile, validatePointByTileMatrix, validateTileMatrixIdByTileMatrixSet } from '../validations';
 import { BaseGeometry } from './baseGeometry';
