+++ src/tiles/tileRange.ts
@@ -4,8 +4,8 @@ import type { ArrayElement } from '../utils/types';
 import { validateMetatile, validateTileMatrixIdByTileMatrixSet } from '../validations';
 import { Tile } from './tile';
 import type { TileMatrixSet } from './tileMatrixSet';
-import { tileMatrixToBBox } from './tiles';
 import type { TileIndex, TileMatrixId, TileMatrixLimits } from './types';
+import { tileMatrixToBBox } from '.';
 
 /**
  * Tile range class that supports a metatile definition
