+++ src/validations/index.ts
@@ -4,9 +4,9 @@ import { SUPPORTED_CRS } from '../constants';
 import type { BoundingBox } from '../geometries/boundingBox';
 import { Point } from '../geometries/point';
 import type { CoordRefSysJSON } from '../geometries/types';
+import { tileMatrixToBBox } from '../tiles';
 import type { TileMatrixSet } from '../tiles/tileMatrixSet';
 import type { TileRange } from '../tiles/tileRange';
-import { tileMatrixToBBox } from '../tiles/tiles';
 import type { CRS as CRSType, TileMatrix, TileMatrixId, TileMatrixSetJSON } from '../tiles/types';
 import type { ArrayElement } from '../utils/types';
