+++ tests/geometries/line.spec.ts
@@ -5,9 +5,9 @@ import type { BBox, Position } from 'geojson';
 import { SUPPORTED_CRS } from '../../src/constants';
 import { Line } from '../../src/geometries/line';
 import type { GeoJSONLineString, LineStringInput } from '../../src/geometries/types';
+import { tileMatrixToBBox } from '../../src/tiles';
 import { TileMatrixSet } from '../../src/tiles/tileMatrixSet';
 import { TILEMATRIXSETJSON_WORLD_CRS84_QUAD } from '../../src/tiles/tileMatrixSets/worldCRS84Quad';
-import { tileMatrixToBBox } from '../../src/tiles/tiles';
 import type { TileMatrixLimits, TileMatrixSetJSON, TileMatrixSet as TileMatrixSetType } from '../../src/tiles/types';
 import { generateLineInput } from './helpers/geometries';
 import { generateNonFinite, isSafeInteger } from './helpers/propertyTest';
