+++ tests/geometries/point.spec.ts
@@ -3,9 +3,9 @@ import { fc, it } from '@fast-check/jest';
 import { strictCircularDeepEqual } from 'fast-equals';
 import { Point } from '../../src/geometries/point';
 import type { GeoJSONPoint, PointInput } from '../../src/geometries/types';
+import { tileMatrixToBBox } from '../../src/tiles';
 import { TileMatrixSet } from '../../src/tiles/tileMatrixSet';
 import { TILEMATRIXSETJSON_WORLD_CRS84_QUAD } from '../../src/tiles/tileMatrixSets/worldCRS84Quad';
-import { tileMatrixToBBox } from '../../src/tiles/tiles';
 import type { TileMatrixLimits, TileMatrixSetJSON, TileMatrixSet as TileMatrixSetType } from '../../src/tiles/types';
 import { generatePointInput } from './helpers/geometries';
 import { generateNonFinite, isSafeInteger } from './helpers/propertyTest';
